_packetListのデータ構造をListからDictionaryに変更し、最新のパケットを参照するように修正。_updateTimeを使用してパケットの更新処理を改善。
